// api/src/CovidTestCenter.ts
export interface CovidTestCenter {
  ID:            string;
  rs:            string; //raison sociale
  adresse:       string;
  do_prel:       string;
  do_antigenic:  string;
  longitude:     string;
  latitude:      string;
  mod_prel:      string;
  public:        string;
  horaire:       string;
  check_rdv:     string;
  tel_rdv:       string;
  web_rdv:       string;
  date_modif:    string;
}